requirements.txt
Flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0